allow universal selectors (` > *`)
We know this selector should be used ligthly, but it is great on some usecases.
When dealing with strict DOM structure like forms or global layout, it can be quite
useful to target elements with selectors like `.ParentElement > *`.

There is no rule to specifically allow `> *` which is the case I think should be allowed.
What should we do then? The straightforward way in this commit is to remove the config.

I think we should allow people to use it