allow consecutive duplicate properties for fallback values
For now the linter doesn't allow us to duplicate properties. We should be able
to in order to write fallbacks values.
For examples, if I want to use viewport related units (*vh*, *vw*), or use *rgba* colors,
and I unfortunately need to support IE8 just because, I need to write alternative values
next to the cool ones.