allow to have 'glued' comments to selectors
the linter warns us if we have code like this:

```scss
// this stuff is super cool
.Stuff {
	font-family: Lato;
}
```

The problem for the linter is that the comment is too close to it. There should
be one empty one between a comment and a non-nested selector.

I think this is too restrictive and we should allow comments like that.